don't set send_notifications for projects copy dependencies
Projects::CopyService wraps everything with service context passing
send_notifications to Journal::NotificationConfiguration, so trying to
set it in dependencies to same value is not needed (and is noop + logs a
warning). Also it is the real reason for
https://community.openproject.org/wp/58342, as Members::CreateService
send_notifications could get set to true and would override
Journal::NotificationConfiguration.active?