Fix flaky spec delete_project_storage_and_file_links_spec.rb:67
failing spec: ./modules/storages/spec/features/delete_project_storage_and_file_links_spec.rb:67
failing run: https://github.com/opf/openproject/actions/runs/18154034691/job/51669959474

What I see when running the spec in headfull mode is that the
`delete_confirmation` input field is filled by Capybara, and then
immediately cleared out by "something". Of course the button is not
enabled because the input field does not match the expected value, and
so the test fails when it tries to click the button.

I could not find what that something is, which would have helped
synchronizing the test with it.Instead I added a 200ms sleep which seems
to have fixed the issue.