[68014] Fix setting milestone date on creation
https://community.openproject.org/wp/68014

When frontend sends the date(s) to the backend, it needs to send `date`
for milestone type instead of `start_date` because the backend would
ignore `startDate` and `dueDate` for milestone type.

Date picker deals with `startDate` and `dueDate` fields only, so the
field `date` is used when calling `handleSuccessfulCreate` to set the
date correctly for milestone types.