Select columns specifically needed for WP autocompletion
This allows to make more efficient database queries in the API
and reduces overhead by too large JSON responses.

Assuming we fix #68457, additional improvements in response time
can be expected, because we do not need to perform a count for all
matching work packages anymore (because we do not select "total").