Fix creation of stories in backlog in dev mode
In development environment, there are some comments added around the
html returned by the server. This makes functions like
`result.$.attr('id')` not work as expected: they'll try to work on the
first child which is a comment, and so will return empty string or
`undefined`.

Fix it by explicitly working on the `.model` element. Also work with
`$(obj.el)` instead of `obj.$` because it's more reliable. Not sure why.

And it can't be covered by feature tests, because in test there are not
these comments. I hope I have not missed something.