Don't allow global memberships on placeholder users
The view_all_principals migration added global roles to all users with
manage_members.

On QA, we had placeholder users with this permission. Placeholder users
should never receive global roles, but there was no validation in place.

This commit adds a migration to remove those, and adds a validation to
the contract.