[68560] Reschedule related work packages even if save occurs during UpdateAncestors
https://community.openproject.org/wp/68560

Previous fix of the same issue was incorrect, because
`UpdateAncestorsService` was no longer saving the work package if the
changes were done outside of the service. Sadly, automatically generated
subjects are updated outside of the service and rely on
`UpdateAncestorsService` to be saved. That's a code smell which should
be fixed in a future refactoring.

An integration test was added to ensure the automatically generated
subject is saved correctly, the fix was reverted, and another fix was
found.

Now the fix is to consider the changes from both the update service and
the `UpdateAncestorsService` when calling the `SetScheduleService`.