Fix error 403 with news widget when news module is disabled
https://community.openproject.org/wp/68850

The widget news controller was checking `:view_news` permission before
rendering. For projects without the news module enabled, the permission
check fails and renders a 403 error.

As the news widget from dashboard tab is rendered through Angular with
`WidgetNewsComponent` using a turbo-frame, the 403 error was somehow
rendering a full page error, preventing the page to be displayed
properly. This prevented people from accessing the project page.

This fixes the issue by disabling the permission check for the news
widget, and displaying a "Nothing new to report." message instead when
the module is disabled.