Add `rm -f db/structure.sql` to reset the test database
When `db/structure.sql` exists, it's used when doing `rails db:migrate`
instead of running the migrations. When switching to a branch with less
migrations, the test database will be created with the
`db/structure.sql` file which contains the migrations of the previous
branch. They should not be there.

To have the correct database structure for tests, the `db/structure.sql`
file must be removed before running `db:migrate`.