Fix SAML single logout
We were already using a few configuration values that work on the 2.x line
of the omniauth-saml gem, but not on version 1.10, which is the one we are still
using.

Since we totally want to migrate to the 2.x line going foward (once we can use omniauth 2.x),
I changed the built configuration to be compatible to both versions of the gem at once,
so that a future upgrade is hopefully smoother.