Fix creating two journals for a new record
When creating a journal, it updates the validity_period of the previous
journal so that the end of the previous journal is the start of the new
journal.

The start of the new journal is determined by the `updated_at` value of
the journable, and this one is updated (touched) in the
`touch_journable` CTE but only if it was not touched by last
modifications. This is checked using
`journable.updated_at_previously_changed?`.

When the journable got just created,

WIP
 If the journable got its `updated_at` changed
recently, then it will use it

Add failing test for adding multiple comments right after creation

Fix the double save issue

Remove debug statement

Fix test