Fix flickering spec
Selection by work package ID was unreliable, because the text
that's selected by is from anywhere in the selected element.

The element shows different information:
* WP subject
* WP ID
* Project name
* status name

The intention of the original code was to uniquely select work packages
by their ID, however it was also perfectly possible to have a project
named "Project 123" as the parent of all work packages. So if the WP
id happens to be 123 as well, the selection would be effectively random
across the work packages.

(Credit for discovering the root cause of this flicker goes to Christophe Bliard
if anyone is asking ;-) )

Fixes flickers of spec/features/work_packages/details/relations/hierarchy_spec.rb,

Reproduction instructions as provided by Christophe:

    rm -f db/structure.sql; \
    bin/rails db:drop db:create db:migrate RAILS_ENV=test > /dev/null ; \
    CI=true TZ=UTC OPENPROJECT_LOG__LEVEL=info CAPYBARA_PUMA_THREADS=0:4 DISABLE_SPRING=1 \
    bin/rspec --seed 34872 -fd \
    'spec/features/work_packages/details/relations/hierarchy_spec.rb[1:1:1:1]' \
    spec/features/work_packages/work_package_index_spec.rb:46